+++ CHANGELOG.md
@@ -6,6 +6,13 @@
  - Bugfix: Clear delayed functions (debounce and wait) on script restart
  - WebWidget: Allow javascript to send `WebWidgetData` via js function `sendData(data)`
  - Bugfix: Events seems to be were out of order judging by their timestamp
+ - Added support for custom events. To generate a custom event: 
+   ```internal:send_custom_event("Hello", { Custom = "Events", Are = "Supported", For = "World"})```
+   These will be published as `InternalCustomEvent` events, having two fields: `Name` and `Json`.
+   Use `parse_json(event.Json)` to get a LuaTable with the values as were sent 
+   via the event. Also adds `generate_json(luaTable)` to generate json string from 
+   a Lua table.
+
 
 ## [0.9.0](https://github.com/dennis/slipstream/releases/tag/v0.9.0) (2021-08-21)
 [Full Changelog](https://github.com/dennis/slipstream/compare/v0.8.0...v0.9.0)